// Coupon System - Front End/src/Models/Category.ts
export enum Category {
    ELECTRICITY = "ELECTRICITY",
    FOOD = "FOOD",
    VACATION = "VACATION",
    HOME_FURNITURE = "HOME_FURNITURE",
    CLOTHING = "CLOTHING",
    TOYS = "TOYS",
    BOOKS = "BOOKS",
    COMPUTERS_AND_PHONES = "COMPUTERS_AND_PHONES",
    SPA_AND_BEAUTY = "SPA_AND_BEAUTY",
}

export const categoryListAll = Object.values(Category);